equalais_rest_server/requirements.txt
absl-py==0.1.13
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
clarifai==2.0.33
click==6.7
configparser==3.5.0
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
enum34==1.1.6
flask-cors==3.0.3
flask==0.12.2
future==0.15.2
gast==0.2.0
git+http://github.com/tensorflow/cleverhans.git#egg=cleverhans
grpcio==1.10.1
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.8.2
ipython-genutils==0.2.0
ipython==6.3.1
ipywidgets==7.2.0
itsdangerous==0.24
jedi==0.11.1
jinja2==2.10
jsonschema==2.5.1
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyter==1.0.0
keras==2.1.5
kiwisolver==1.0.1
markdown==2.6.11
markupsafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0; sys_platform != 'win32'
pickleshare==0.7.4
pillow==2.9.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.4.3
ptyprocess==0.5.2; os_name != 'nt'
pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
pywavelets==0.5.2
pyyaml==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.13.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.1
send2trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
wcwidth==0.1.7
werkzeug==0.14.1
wheel==0.31.0
widgetsnbextension==3.2.0
